Move styles folder into static folder
