+++ client/components/LoginForm.js
@@ -3,7 +3,7 @@ import Link from "next/link";
 import {Form, Input, Button, Icon} from "antd";
 
 import "antd/dist/antd.css";
-import "../styles/LoginForm.css";
+import "../static/styles/LoginForm.css";
 
 const FormItem = Form.Item;
