+++ client/pages/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 
 import Login from "./login";
-import "../styles/index.css"; // eslint-disable-line import/no-unassigned-import
+import "../static/styles/index.css"; // eslint-disable-line import/no-unassigned-import
 
 export default () => <Login/>;
