+++ client/static/styles/Footer.css
@@ -1,0 +1,8 @@
+.footer {
+    text-align: center;
+    height: 2rem; 
+    background: #333 !important; 
+    color: #ddd !important;
+    display: grid;
+    align-content: center;
+}
