+++ client/static/styles/RegisterForm.css
@@ -1,0 +1,28 @@
+.register__form {
+    display: grid;
+    grid-gap: 10px;
+    text-align: center;
+    padding: 1rem;
+    margin: 0;
+    padding: 0;
+}
+
+.register__form > * {
+    margin: 0;
+}
+
+.register__form button {
+    margin-top: 0.75rem;
+}
+
+
+.register__form--title {
+    text-align: center;
+}
+
+@media(min-width: 30rem) {
+    .register__form {
+        justify-content: center;
+        align-content: center;
+    }
+}
