+++ client/static/styles/Sidebar.css
@@ -1,0 +1,10 @@
+.logo {
+    background: url(/static/logo.png);
+    background-repeat: no-repeat;
+    background-size: cover;
+    width: 6rem;
+    height: 6rem;
+    margin: auto;
+    margin-top: 1.5rem;
+    margin-bottom: 0.5rem;
+}
